+++ app/builders/pathfinder2_character/classes/druid_builder.rb
@@ -18,6 +18,7 @@ module Pathfinder2Character
         result[:spell_attack] = 1
         result[:spell_list] = 'primal'
         result[:main_ability] = 'wis'
+        result[:feats] = result[:feats].push('shield_block').uniq
 
         result
       end