+++ app/builders/pathfinder2_character/subclasses/leaf_builder.rb
@@ -5,6 +5,8 @@ module Pathfinder2Character
     class LeafBuilder
       def call(result:)
         result[:skill_boosts].merge!({ diplomacy: 1 }) { |_, oldval, newval| oldval + newval }
+        result[:feats] = result[:feats].push('leshy_familiar').uniq
+        result[:focus_spells] = result[:focus_spells].push('cornucopia').uniq
 
         result
       end