+++ app/builders/pathfinder2_character/subclasses/silence_in_snow_builder.rb
@@ -6,6 +6,8 @@ module Pathfinder2Character
       def call(result:)
         result[:skill_boosts].merge!({ nature: 1 }) { |_, oldval, newval| oldval + newval }
         result[:spell_list] = 'primal'
+        result[:focus_spells] = result[:focus_spells].push('clinging_ice').uniq
+        result[:spells] = result[:spells].push('gust_of_wind').uniq
 
         result
       end