+++ app/builders/pathfinder2_character/subclasses/spinner_of_threads_builder.rb
@@ -6,6 +6,8 @@ module Pathfinder2Character
       def call(result:)
         result[:skill_boosts].merge!({ occultism: 1 }) { |_, oldval, newval| oldval + newval }
         result[:spell_list] = 'occult'
+        result[:focus_spells] = result[:focus_spells].push('nudge_fate').uniq
+        result[:spells] = result[:spells].push('sure_strike').uniq
 
         result
       end