+++ app/builders/pathfinder2_character/subclasses/wilding_steward_builder.rb
@@ -6,6 +6,7 @@ module Pathfinder2Character
       def call(result:)
         result[:skill_boosts].merge!({ nature: 1 }) { |_, oldval, newval| oldval + newval }
         result[:spell_list] = 'primal'
+        result[:focus_spells] = result[:focus_spells].push('wilding_word').uniq
 
         result
       end